requirements.txt
# ============================================
# MAILFLOW — Python Dependencies
# ============================================
flask>=3.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
torch>=2.0.0
transformers>=4.30.0
requests>=2.31.0
google-generativeai>=0.5.0
